Revert "chore: Rename sidebar item injectable files"
This reverts commit 72a1e2aaa0d67d3489af292df21cb29a4c6a5101.
