+++ packages/core/src/renderer/components/cluster/cluster-overview-sidebar-items.injectable.tsx
