+++ packages/core/src/renderer/components/config/config-sidebar-items.injectable.tsx
