+++ packages/core/src/renderer/components/custom-resources/custom-resource-definition-groups-sidebar-items-computed.injectable.ts
@@ -14,7 +14,7 @@ import navigateToCustomResourcesInjectable from "../../../common/front-end-routi
 import { shouldShowResourceInjectionToken } from "../../../features/cluster/showing-kube-resources/common/allowed-resources-injection-token";
 import routeIsActiveInjectable from "../../routes/route-is-active.injectable";
 import routePathParametersInjectable from "../../routes/route-path-parameters.injectable";
-import customResourcesSidebarItemInjectable from "./custom-resources-sidebar-item.injectable";
+import customResourcesSidebarItemInjectable from "./custom-resources-sidebar-items.injectable";
 import customResourceDefinitionsInjectable from "./custom-resources.injectable";
 
 const customResourceDefinitionGroupsSidebarItemsComputedInjectable = getInjectable({
