+++ packages/core/src/renderer/components/custom-resources/custom-resources-sidebar-items.injectable.tsx
