+++ packages/core/src/renderer/components/events/events-sidebar-items.injectable.tsx
