+++ packages/core/src/renderer/components/helm/helm-sidebar-items.injectable.tsx
