+++ packages/core/src/renderer/components/namespaces/namespaces-sidebar-items.injectable.tsx
