+++ packages/core/src/renderer/components/network-ingresses/ingresses-sidebar-items.injectable.ts
@@ -7,7 +7,7 @@ import { getInjectable } from "@ogre-tools/injectable";
 import navigateToIngressesInjectable from "../../../common/front-end-routing/routes/cluster/network/ingress-class/navigate-to-ingress-classes.injectable";
 import ingressesRouteInjectable from "../../../common/front-end-routing/routes/cluster/network/ingresses/ingresses-route.injectable";
 import routeIsActiveInjectable from "../../routes/route-is-active.injectable";
-import networkSidebarItemInjectable from "../network/network-sidebar-item.injectable";
+import networkSidebarItemInjectable from "../network/network-sidebar-items.injectable";
 
 const ingressesSidebarItemInjectable = getInjectable({
   id: "ingresses-sidebar-item",
