+++ packages/core/src/renderer/components/network/network-sidebar-items.injectable.tsx
