+++ packages/core/src/renderer/components/nodes/nodes-sidebar-items.injectable.tsx
