+++ packages/core/src/renderer/components/user-management/user-management-sidebar-items.injectable.tsx
