+++ packages/core/src/renderer/components/workloads/workloads-sidebar-items.injectable.tsx
